feat(send-and-read-emails): merge node examples
