+++ packages/send-and-read-emails/backend/node/route.js
@@ -1,43 +1,53 @@
 const { default: Draft } = require('nylas/lib/models/draft');
-const mockDb = require('./utils/mock-db');
-const { getReqBody } = require('./utils/mock-server');
 
 exports.sendEmail = async (req, res, nylasClient) => {
-  if (!req.headers.authorization) {
-    return res.writeHead(401).end('Unauthorized');
-  }
+  const user = res.locals.user;
 
-  const user = await mockDb.findUser(req.headers.authorization);
-  if (!user) {
-    return res.writeHead(401).end('Unauthorized');
-  }
-
-  const { to, body } = await getReqBody(req);
+  const { to, subject, body, replyToMessageId } = req.body;
 
   const draft = new Draft(nylasClient.with(user.accessToken));
 
   draft.from = [{ email: user.emailAddress }];
   draft.to = [{ email: to }];
+  draft.subject = subject;
   draft.body = body;
+  draft.replyToMessageId = replyToMessageId;
 
   const message = await draft.send();
 
-  return res.writeHead(200).end(JSON.stringify(message));
+  return res.json(message);
 };
 
 exports.readEmails = async (req, res, nylasClient) => {
-  if (!req.headers.authorization) {
-    return res.writeHead(401).end('Unauthorized');
-  }
-
-  const user = await mockDb.findUser(req.headers.authorization);
-  if (!user) {
-    return res.writeHead(401).end('Unauthorized');
-  }
+  const user = res.locals.user;
 
   const nylas = nylasClient.with(user.accessToken);
 
-  const threads = await nylas.threads.list({ limit: 5 });
+  const threads = await nylas.threads.list({ limit: 50, expanded: true });
 
-  return res.writeHead(200).end(JSON.stringify({ threads }));
+  return res.json(threads);
+};
+
+exports.getMessage = async (req, res, nylasClient) => {
+  const user = res.locals.user;
+
+  const nylas = nylasClient.with(user.accessToken);
+
+  const { id } = req.query;
+  const message = await nylas.messages.find(id);
+
+  return res.json(message);
+};
+
+exports.getFile = async (req, res, nylasClient) => {
+  const user = res.locals.user;
+
+  const nylas = nylasClient.with(user.accessToken);
+
+  const { id } = req.query;
+  const file = await nylas.files.find(id);
+
+  // Files will be returned as a binary object
+  const fileData = await file.download();
+  return res.end(fileData?.body);
 };
