+++ packages/send-and-read-emails/backend/node/server.js
@@ -1,13 +1,12 @@
+const express = require('express');
+const cors = require('cors');
 const dotenv = require('dotenv');
 const mockDb = require('./utils/mock-db');
-const { mockServer, getReqBody } = require('./utils/mock-server');
 const route = require('./route');
 
 const Nylas = require('nylas');
 const { WebhookTriggers } = require('nylas/lib/models/webhook');
 const { Scope } = require('nylas/lib/models/connect');
-const { Routes: NylasRoutes } = require('nylas/lib/services/routes');
-const { DefaultPaths } = require('nylas/lib/services/routes');
 const { openWebhookTunnel } = require('nylas/lib/services/tunnel');
 
 dotenv.config();
@@ -21,58 +20,53 @@ const nylasClient = new Nylas({
   clientSecret: process.env.NYLAS_CLIENT_SECRET,
 });
 
+const app = express();
+
+// Enable CORS
+app.use(cors());
+
 // The uri for the frontend
 const CLIENT_URI =
   process.env.CLIENT_URI || `http://localhost:${process.env.PORT || 3000}`;
 
-// Use the routes provided by the Nylas Node SDK to quickly implement
-// the authentication flow
-const { buildAuthUrl, exchangeCodeForToken } = NylasRoutes(nylasClient);
+// '/nylas/generate-auth-url': This route builds the URL for
+// authenticating users to your Nylas application via Hosted Authentication
+app.post('/nylas/generate-auth-url', express.json(), async (req, res) => {
+  const { body } = req;
 
-// Configure the Nylas routes using your flavour of backend framework
-// '/nylas/generate-auth-url': This route builds the URL for authenticating
-// users to your Nylas application via Hosted Authentication
-mockServer.post(DefaultPaths.buildAuthUrl, async (req, res) => {
-  const body = await getReqBody(req);
-
-  const authUrl = await buildAuthUrl({
-    scopes: [Scope.EmailModify, Scope.EmailSend, Scope.EmailReadOnly],
-    emailAddress: body.email_address,
-    successUrl: body.success_url,
-    clientUri: CLIENT_URI,
+  const authUrl = nylasClient.urlForAuthentication({
+    loginHint: body.email_address,
+    redirectURI: (CLIENT_URI || '') + body.success_url,
+    scopes: [Scope.EmailReadOnly, Scope.EmailModify, Scope.EmailSend],
   });
 
-  res.writeHead(200).end(authUrl);
+  return res.send(authUrl);
 });
 
 // '/nylas/exchange-mailbox-token': This route exchanges an authorization
 // code for an access token
-mockServer.post(DefaultPaths.exchangeCodeForToken, async (req, res) => {
-  const body = await getReqBody(req);
+// and sends the details of the authenticated user to the client
+app.post('/nylas/exchange-mailbox-token', express.json(), async (req, res) => {
+  const body = req.body;
 
-  try {
-    const { accessToken, emailAddress } = await exchangeCodeForToken(
-      body.token
-    );
+  const { accessToken, emailAddress } = await nylasClient.exchangeCodeForToken(
+    body.token
+  );
 
-    // Normally store the access token in the DB
-    console.log('Access Token was generated for: ' + emailAddress);
-    // Replace this mock code with your actual database operations
-    const user = await mockDb.createOrUpdateUser(emailAddress, {
-      accessToken,
-      emailAddress,
-    });
+  // Normally store the access token in the DB
+  console.log('Access Token was generated for: ' + emailAddress);
 
-    // Return an authorization object to the user
-    res.writeHead(200).end(
-      JSON.stringify({
-        id: user.id,
-        emailAddress: user.emailAddress,
-      })
-    );
-  } catch (e) {
-    res.writeHead(500).end(e.message);
-  }
+  // Replace this mock code with your actual database operations
+  const user = await mockDb.createOrUpdateUser(emailAddress, {
+    accessToken,
+    emailAddress,
+  });
+
+  // Return an authorization object to the user
+  return res.json({
+    id: user.id,
+    emailAddress: user.emailAddress,
+  });
 });
 
 // Start the Nylas webhook
@@ -86,20 +80,53 @@ openWebhookTunnel(nylasClient, {
           JSON.stringify(delta.objectData, undefined, 2)
         );
         break;
+      case WebhookTriggers.AccountConnected:
+        console.log(
+          'Webhook trigger received, account connected. Details: ',
+          JSON.stringify(delta.objectData, undefined, 2)
+        );
+        break;
     }
   },
-}).then((webhookDetails) =>
-  console.log('Webhook tunnel registered. Webhook ID: ' + webhookDetails.id)
-);
+}).then((webhookDetails) => {
+  console.log('Webhook tunnel registered. Webhook ID: ' + webhookDetails.id);
+});
+
+// Middleware to check if the user is authenticated
+async function isAuthenticated(req, res, next) {
+  if (!req.headers.authorization) {
+    return res.status(401).json('Unauthorized');
+  }
+
+  // Query our mock db to retrieve the stored user access token
+  const user = await mockDb.findUser(req.headers.authorization);
+
+  if (!user) {
+    return res.status(401).json('Unauthorized');
+  }
+
+  // Add the user to the response locals
+  res.locals.user = user;
+
+  next();
+}
 
 // Handle routes
-mockServer.post('/nylas/send-email', (req, res) =>
+app.post('/nylas/send-email', isAuthenticated, express.json(), (req, res) =>
   route.sendEmail(req, res, nylasClient)
 );
 
-mockServer.get('/nylas/read-emails', (req, res) =>
+app.get('/nylas/read-emails', isAuthenticated, (req, res) =>
   route.readEmails(req, res, nylasClient)
 );
+
+app.get('/nylas/message', isAuthenticated, async (req, res) => {
+  route.getMessage(req, res, nylasClient);
+});
+
+app.get('/nylas/file', isAuthenticated, async (req, res) => {
+  route.getFile(req, res, nylasClient);
+});
 
 // Before we start our backend, we should whitelist our frontend as a redirect
 // URI to ensure the auth completes
@@ -110,10 +137,9 @@ nylasClient
   .then((applicationDetails) => {
     console.log(
       'Application whitelisted. Application Details: ',
-      JSON.stringify(applicationDetails, undefined, 2)
+      JSON.stringify(applicationDetails)
     );
   });
 
 // Start listening on port 9000
-mockServer.init().listen(port);
-console.log('App listening on port ' + port);
+app.listen(port, () => console.log('App listening on port ' + port));
